combined auth and accountServer anded favico and title to index.html
