+++ bank/public/index.html
@@ -2,14 +2,13 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+    <link rel="icon" href="./favicon.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
-      name="description"
-      content=""
+      name="Dice Bank"
     />
-    <title>Bank</title>
+    <title>Dice Bank</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
