+++ bank/server/controllers/signUp/isOptionsConflict.js
