+++ bank/server/controllers/signUp/setUpPayments.js
