+++ bank/src/api/axios.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 export default axios.create({
-  baseURL : 'http://localhost:9999/'
+  baseURL : 'http://localhost:7777/'
 })
 
 export const axiosPrivate = axios.create({
